+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Features
 * **core:** Add permission field for project listing
+* **core:** Add `timezone` column to `streams` table([CE-1168](https://jira.rfcx.org/browse/CE-1168))
 * **core:** Segment endpoint supporting prediction service to get a whole segment file efficiently ([CE-1166](https://jira.rfcx.org/browse/CE-1166))
 * **core:** Standardised segment endpoint security